Fix crash when reentering cluster map fragment. Move duplicate code to superclass. Increase build version.
