+++ app/build.gradle
@@ -8,8 +8,8 @@ android {
         applicationId "dk.hug.treehugger"
         minSdkVersion 19
         targetSdkVersion 23
-        versionCode 1
-        versionName "1.0"
+        versionCode 3
+        versionName "1.2"
     }
 
     buildTypes {
